# Use torch.repeat_interleave() to generate the repeated indices faster